Rykket koebFelt over i Handlinger, samt fjernet gammel klasse
