+++ src/main/java/controller/subcontroller/HandelHotel.java
@@ -1,0 +1,7 @@
+package controller.subcontroller;
+
+public class HandelHotel {
+
+
+
+}
