+++ src/main/java/controller/subcontroller/HandelHus.java
@@ -1,0 +1,8 @@
+package controller.subcontroller;
+
+public class HandelHus {
+
+
+
+
+}
