+++ src/main/java/controller/subcontroller/Handlinger.java
@@ -1,6 +1,7 @@
 package controller.subcontroller;
 
 import controller.SpilController;
+import model.felter.ejeligefelter.EjeligtFelt;
 import model.raflebaeger.RafleBaeger;
 import view.UserInterface;
 import model.Spiller;
@@ -29,6 +30,21 @@ public class Handlinger {
         }
     }
 
+
+    public void koebFelt(EjeligtFelt felt, Spiller spiller, UserInterface ui ){
+
+        if (spiller.getPenge() >  felt.getPris()) {
+
+            spiller.addPenge( -  felt.getPris() );
+            felt.setEjer(spiller);
+            spiller.addEjeligtFelt( felt );
+
+            ui.gennemfoertKoeb(  felt, spiller);
+
+        } else {
+            ui.manglerPenge();
+        }
+    }
 
 
     /**
